+++ ui/image_factory.go
@@ -20,8 +20,8 @@ func (f *imageFactory) CreateImage(img image.Image) *Image {
 	bounds := img.Bounds()
 	size := NewSize(bounds.Dx(), bounds.Dy())
 	texture := f.api.CreateColorTexture2D(render.ColorTexture2DInfo{
-		Width:           size.Width,
-		Height:          size.Height,
+		Width:           uint32(size.Width),
+		Height:          uint32(size.Height),
 		GenerateMipmaps: true,
 		GammaCorrection: false,
 		Format:          render.DataFormatRGBA8,
